move fit() to BaseForecast class
