+++ feature_engine/timeseries/forecasting/__init__.py
@@ -1,9 +1,11 @@
 """ Transformers that create features for time-series forecasting."""
 
+from .base_forecast import BaseForecast
 from .lag_features import LagFeatures
 from .window_features import WindowFeatures
 
 __all__ = [
+    "BaseForecast",
     "LagFeatures",
     "WindowFeatures",
 ]
